Adding synthesize.ts to orchestrate chart synthesis
