+++ package.json
@@ -10,6 +10,7 @@
     "dedent-js": "^1.0.1",
     "fast-glob": "^3.3.2",
     "find-root": "^1.1.0",
+    "js-yaml": "^4.1.0",
     "json-stable-stringify": "^1.3.0",
     "yaml": "^2.3.4"
   },
@@ -18,6 +19,7 @@
     "@types/common-tags": "^1.8.4",
     "@types/debug": "^4.1.12",
     "@types/find-root": "^1.1.4",
+    "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.0.0",
     "@typescript-eslint/eslint-plugin": "^8.0.0-alpha.14",
     "@typescript-eslint/parser": "^8.0.0-alpha.14",
